function/requirements.txt
azure-functions
azure-functions-durable>=1.0.0
requests
langgraph>=0.0.40
langchain-core>=0.1.0
reportlab>=4.0.0
PyPDF2>=3.0.0
nltk>=3.8
textstat>=0.7.0
weasyprint>=60.0
